refactor common to lib
