+++ resources/haytack_tags/haystack-3.0.1/bin/fanlaunch
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# fanlaunch: common code for unix scripts
+# fanlaunch: lib code for unix scripts
 #
 
 # Operating system
