+++ worker/add_site/add_osw/add_osw.py
@@ -33,9 +33,9 @@ import zipfile
 import shutil
 from subprocess import call
 import logging
-import common
+import lib
 
-(osw_zip_name, upload_id, directory) = common.precheck_argus(sys.argv)
+(osw_zip_name, upload_id, directory) = lib.precheck_argus(sys.argv)
 
 logger = logging.getLogger('addsite')
 logger.setLevel(logging.INFO)
@@ -96,9 +96,9 @@ shutil.copyfile(user_epwpath, epwpath)
 
 call(['openstudio', 'run', '-m', '-w', workflowpath])
 
-common.make_ids_unique(upload_id, points_jsonpath, mapping_jsonpath)
-common.replace_siteid(upload_id, points_jsonpath, mapping_jsonpath)
+lib.make_ids_unique(upload_id, points_jsonpath, mapping_jsonpath)
+lib.replace_siteid(upload_id, points_jsonpath, mapping_jsonpath)
 
-common.upload_site_DB_Cloud(points_jsonpath, bucket, directory)
+lib.upload_site_DB_Cloud(points_jsonpath, bucket, directory)
 
 shutil.rmtree(directory)
