+++ worker/lib/data/categories.json
